api/requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
constants==0.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
h11==0.9.0
httptools==0.1.1
idna==2.9
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
lml==0.0.9
MarkupSafe==1.1.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.3
pip-review==1.1.0
psycopg2-binary==2.8.4
pycparser==2.20
pyparsing==2.4.6
python-magic==0.4.15
pytz==2019.3
requests==2.23.0
serializers==0.2.4
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
texttable==1.6.2
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
